requirements.txt
streamlit
groq
sentence-transformers
faiss-cpu
pypdf
python-docx
youtube-transcript-api
duckduckgo-search
pandas
python-dotenv
numpy
